+++ fiddles/angular2/fiddle-0029-Dashboard/tools/config.js
@@ -1,4 +1,0 @@
-var project_config_1 = require('./config/project.config');
-var config = new project_config_1.ProjectConfig();
-module.exports = config;
-//# sourceMappingURL=config.js.map