+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/project/build.fonts.js
@@ -1,7 +1,0 @@
-var gulp = require('gulp');
-var config_1 = require('../../config');
-module.exports = function () {
-    return gulp.src(config_1.FONTS_SRC)
-        .pipe(gulp.dest(config_1.FONTS_DEST));
-};
-//# sourceMappingURL=build.fonts.js.map