+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/seed/clean.all.js
@@ -1,4 +1,0 @@
-var config_1 = require('../../config');
-var utils_1 = require('../../utils');
-module.exports = utils_1.clean(config_1.DIST_DIR);
-//# sourceMappingURL=clean.all.js.map