+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/seed/clean.prod.js
@@ -1,4 +1,0 @@
-var config_1 = require('../../config');
-var utils_1 = require('../../utils');
-module.exports = utils_1.clean([config_1.PROD_DEST, config_1.TMP_DIR]);
-//# sourceMappingURL=clean.prod.js.map