+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/seed/karma.start.js
@@ -1,9 +1,0 @@
-var karma = require('karma');
-var path_1 = require('path');
-module.exports = function (done) {
-    new karma.Server({
-        configFile: path_1.join(process.cwd(), 'karma.conf.js'),
-        singleRun: true
-    }).start(done);
-};
-//# sourceMappingURL=karma.start.js.map