+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/seed/server.start.js
@@ -1,3 +1,0 @@
-var utils_1 = require('../../utils');
-module.exports = utils_1.serveSPA;
-//# sourceMappingURL=server.start.js.map