+++ fiddles/angular2/fiddle-0029-Dashboard/tools/tasks/seed/watch.dev.js
@@ -1,3 +1,0 @@
-var utils_1 = require('../../utils');
-module.exports = utils_1.watch('build.dev');
-//# sourceMappingURL=watch.dev.js.map