+++ fiddles/angular2/fiddle-0029-Dashboard/tools/utils/project/sample_util.js
@@ -1,4 +1,0 @@
-function myUtil() {
-}
-exports.myUtil = myUtil;
-//# sourceMappingURL=sample_util.js.map