+++ fiddles/angular2/fiddle-0029-Dashboard/tools/utils/seed/template_locals.js
@@ -1,6 +1,0 @@
-var CONFIG = require('../../config');
-function templateLocals() {
-    return CONFIG;
-}
-exports.templateLocals = templateLocals;
-//# sourceMappingURL=template_locals.js.map