+++ fiddles/d3/fiddle-0017-ClickEventListener/app.js
@@ -39,3 +39,4 @@
         }
     });
 })(window.app = window.app || {}, d3)
+d